requirements.txt
Pillow
RPi.GPIO
paho-mqtt
customtkinter